refactor(next-core): extract custom transforms
